+++ src/components/Card/Card.css
@@ -8,3 +8,5 @@
     border-radius: 0.25rem;
     box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.15);
 }
+
+
